requirements.txt
matplotlib>=3.5.0
numpy>=1.21.0
jieba>=0.42.1
PyYAML>=6.0
chardet>=5.0.0
